requirements.txt
# Django y REST Framework
Django==3.2.24
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caché y colas
redis==5.0.1
celery==5.3.6

# Almacenamiento
boto3==1.34.14
django-storages==1.14.2

# Machine Learning
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

# Utilidades
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# PDF generation
markdown2pdf==0.1.0

# Added from the code block
reportlab==4.1.0

